middleware pipeline for api search

+++ controllers/issueController.js
@@ -4,11 +4,12 @@ const mongoose = require('mongoose');
 const Issue = mongoose.model('Issue');
 
 exports.httpsRequest = (req, res, next) => {
-  const jql = (req.query.jql !== {}) ? req.query.jql : 'filter=22119';
+  // const jql = (req.query.jql !== {}) ? req.query.jql : 'filter=22119';
+  const jql = 'filter=22119';
   const bodyData = JSON.stringify({
     jql,
     startAt: 0,
-    maxResults: 500,
+    maxResults: 200,
     fields: [
       "summary",
       "status",
@@ -56,64 +57,41 @@ exports.httpsRequest = (req, res, next) => {
   request.end();
 }
 
-exports.shallowCopy= async (req, res, next) => {
-  const shallowIssues = req.issues.map(({ id, key, fields }) => 
-    ({
-      id,
-      key,
-      assignee: fields.assignee && fields.assignee.key,
-      avatarUrls: fields.assignee && fields.assignee.avatarUrls,
-      summary: fields.summary,
-      issuetype: fields.issuetype.name,
-      status: fields.status.name,
-      statusCategory: fields.status.statusCategory.key,
-      priority: fields.priority.name,
-      components: fields.components,
-    })
-  );
+exports.shallowCopy = async (req, res, next) => {
+  if (req.issues) {
+    console.log(`Saving ${req.issues.length} to MongoDB...`);
+    const shallowIssues = req.issues.map(({ id, key, fields }) =>
+      ({
+        id,
+        key,
+        assignee: fields.assignee.key,
+        summary: fields.summary,
+        issuetype: fields.issuetype.name,
+        status: fields.status.name,
+        statusCategory: fields.status.statusCategory.key,
+        priority: fields.priority.name,
+        components: fields.components,
+      })
+    );
 
-  await Issue.deleteMany();
-  
-  try {
-    await Issue.insertMany(shallowIssues);
+    await Issue.deleteMany();
+
+    try {
+      await Issue.insertMany(shallowIssues);
+      console.log(`Successfully saved ${req.issues.length} to MongoDB!`);
+      next();
+    } catch (e) {
+      console.error(e);
+    }
+  } else {
+    console.log(`Skip saving to MongoDB`);
     next();
-  } catch (e) {
-    console.error(e);
   }
 }
 
-exports.getIssues = async (req, res) => res.json(await Issue.find());
+exports.getIssues = async (req, res, next) => {
+  console.log('Returning issues from MongoDB collection')
+  return res.json(await Issue.find());
+}
 
-exports.getQuery =  async (req, res) => res.json(req.query.jql);
-
-
-
-// HTTPS POST PROMISE
-//
-// return new Promise((resolve, reject) => {
-//   const req = https.request(options, (res) => {
-//     console.log(`STATUS: ${res.statusCode}`);
-//     console.log(`HEADERS: ${JSON.stringify(res.headers)}`);
-//     let rawData = '';
-//     res.setEncoding('utf8');
-//     res.on('data', (chunk) => {
-//       rawData += chunk;
-//     });
-//     res.on('end', () => {
-//       try {
-//         const response = JSON.parse(rawData);
-//         resolve(response);
-//       } catch (err) {
-//         console.error(err.message);
-//       }
-//     });
-//   });
-
-//   req.on('error', (e) => {
-//     console.error(`problem with request: ${e.message}`);
-//     reject();
-//   })
-
-//   req.write(bodyData)
-//   req.end();
-// });
+exports.getQuery = async (req, res) => res.json(req.query.jql);
